refactor(모달 컴포넌트) : 위치 변경에 따른 경로 수정
